Don't mind the bomb

+++ cleanup.py
@@ -19,8 +19,12 @@ for i in list:
 list = os.listdir(".")
 for i in list:
     if i.endswith(".dat"):
-        dir = i.replace(".dat", "")
+        stage = step.parseSteps(i.replace(".dat", ""))
+        files = [re.sub(".*?\\\\", "", s.imageName) for s in stage.steps if s.imageName is not None]
+        dir = stage.name
+        print(files)
+        print(dir)
         allFiles = os.listdir(dir + "\\")
         for f in allFiles:
-            Image.open(dir + "\\" + f).save(dir+"\\" + f.replace(".bmp", ".png"), "png")
-            os.remove(dir + "\\" + f)
+            if f not in files:
+                os.remove(dir + "\\" + f)
